refactor: split types into dedicated files
