+++ Sources/TranscriptionError.swift
@@ -49,28 +49,3 @@ extension TranscriptionError: LocalizedError {
     }
 }
 
-/// A unit of audio data independent from platform frameworks.
-struct AudioChunk: Sendable, Equatable {
-    public var bytes: Data
-    public var sampleRate: Double
-    public var channels: Int
-    public var isFinal: Bool
-
-    public init(bytes: Data, sampleRate: Double, channels: Int, isFinal: Bool = false) {
-        self.bytes = bytes
-        self.sampleRate = sampleRate
-        self.channels = channels
-        self.isFinal = isFinal
-    }
-}
-
-/// Transcription result carrying interim and final texts.
-struct TranscriptionResult: Sendable, Equatable {
-    public let text: String
-    public let isFinal: Bool
-
-    public init(text: String, isFinal: Bool) {
-        self.text = text
-        self.isFinal = isFinal
-    }
-}
