+++ Sources/TranscriptionResult.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+/// Transcription result carrying interim and final texts.
+struct TranscriptionResult: Sendable, Equatable {
+    public let text: String
+    public let isFinal: Bool
+
+    public init(text: String, isFinal: Bool) {
+        self.text = text
+        self.isFinal = isFinal
+    }
+}
+
